requirements.txt
spade
python-dotenv
numpy
pandas
tensorflow
scikit-learn
keras
matplotlib
scipy
spyder
spyder-kernels
cvxopt
PICOS==1.1.2
pandapower
pypsa